totalsales

+++ Backend/mysite/api/views_dir/product_view.py
@@ -3,7 +3,7 @@ from django.shortcuts import get_object_or_404
 from rest_framework import status
 from rest_framework.views import APIView, Response
 
-from ..models import Product, ProductCategory
+from ..models import Product, ProductCategory,ItemActivity
 from ..serializer_dir.item_activity_serializer import ItemActivitySerializer
 from ..serializer_dir.product_serializer import ProductSerializer
 
@@ -303,6 +303,17 @@ class ProductViewClass(APIView):
 
         if serializer.is_valid():
             updated_product = serializer.save()
+            itemactivity = {
+                        "change": updated_product.product_quantity,
+                        "quantity": updated_product.product_quantity,
+                        "product": updated_product.id,
+                        "types": "EDIT_STOCK",
+                        "remarks": "Edit Stock",
+                    }
+
+            itemserializer = ItemActivitySerializer(data=itemactivity)
+            if itemserializer.is_valid():
+                    itemserializer.save()
 
             # Get new data for audit
             new_data = {
@@ -330,6 +341,7 @@ class ProductViewClass(APIView):
                     "message": "Product updated successfully",
                     "data": serializer.data,
                     "changes": {"old": old_data, "new": new_data},
+                    "item_activity":itemserializer.data
                 },
                 status=status.HTTP_200_OK,
             )
@@ -368,21 +380,28 @@ class ProductViewClass(APIView):
                 product = Product.objects.get(id=id, category__branch=my_branch)
 
             product_name = product.name
-            try:
-                product.delete()
-            except Exception:
-                return Response(
-                    {"success": False, "message": "Cannot delete product."},
-                    status=status.HTTP_404_NOT_FOUND,
-                )
+            activity = product.to_product.all()
+    
+            if activity.count() == 1 and activity.filter(remarks = "Opening Stock").exists(): 
+                with transaction.atomic():       
+                    activity.delete()
+                    product.delete()
 
+                return Response(
+                        {
+                            "success": True,
+                            "message": f"Product '{product_name}' deleted successfully.",
+                        },
+                        status=status.HTTP_200_OK,
+                    )
+        
+            
             return Response(
-                {
-                    "success": True,
-                    "message": f"Product '{product_name}' deleted successfully.",
-                },
-                status=status.HTTP_200_OK,
+                {"success": False, "message": "Cannot delete product."},
+                status=status.HTTP_404_NOT_FOUND,
             )
+
+            
 
         except Product.DoesNotExist:
             return Response(
